Switch to TypeScript
