+++ .gitignore
@@ -1,1 +1,11 @@
-node_modules/
+node_modules
+.node_repl_history
+.config
+.npm
+.idea
+.DS_Store
+
+**/*.js
+
+**/*.js.map
+*.d.ts
