+++ .prettierrc
@@ -1,0 +1,5 @@
+{
+	"trailingComma": "es5",
+	"tabWidth": 4,
+	"useTabs": true
+}
