+++ lib/util/to-string.ts
@@ -1,0 +1,8 @@
+export default function toString(val: any) {
+	if (Array.isArray(val)) {
+		return val.join("");
+	} else if (val == null) {
+		return "";
+	}
+	return val.toString();
+}
