+++ package.json
@@ -1,24 +1,28 @@
 {
-  "name": "flora-tmpl",
-  "version": "1.0.0",
-  "description": "Streaming templates for node",
-  "main": "lib/index.js",
-  "scripts": {
-    "test": "mocha test/test.js",
-    "debug": "node --inspect-brk node_modules/.bin/_mocha test/test.js"
-  },
-  "keywords": [
-    "templates"
-  ],
-  "author": "Matthew Phillips",
-  "license": "BSD-2-Clause",
-  "dependencies": {
-    "is-promise": "^2.1.0",
-    "is-stream": "^1.1.0",
-    "stream-array": "^1.1.2"
-  },
-  "devDependencies": {
-    "mocha": "^3.1.0",
-    "through2": "^2.0.3"
-  }
+	"name": "flora-tmpl-ts",
+	"version": "1.0.0",
+	"description": "Streaming templates for node",
+	"main": "lib/index.js",
+	"types": "lib/index.d.ts",
+	"scripts": {
+		"test": "mocha test/test.js",
+		"debug": "node --inspect-brk node_modules/.bin/_mocha test/test.js",
+		"build": "tsc",
+		"watch": "tsc --watch"
+	},
+	"keywords": ["templates"],
+	"author": "Matthew Phillips",
+	"license": "BSD-2-Clause",
+	"dependencies": {
+		"@types/stream-array": "^1.1.0",
+		"is-promise": "^2.1.0",
+		"is-stream": "^2.0.0",
+		"stream-array": "^1.1.2"
+	},
+	"devDependencies": {
+		"@types/node": "^14.11.8",
+		"mocha": "^3.1.0",
+		"through2": "^2.0.3",
+		"typescript": "^4.0.3"
+	}
 }
